v2.9.4.1

+++ 400_xowa/src/gplx/xowa/parsers/xndes/Xop_xatr_whitelist_mgr_tst.java
@@ -49,12 +49,12 @@ class Xop_xatr_whitelist_fxt {
 	public void Clear() {
 		if (whitelist_mgr == null) whitelist_mgr = new Xop_xatr_whitelist_mgr().Ini();
 	}	private Xop_xatr_whitelist_mgr whitelist_mgr;
-	public void Whitelist(byte tag_id, String key_str, boolean expd) {
+	public void Whitelist(int tag_id, String key_str, boolean expd) {
 		byte[] key_bry = Bry_.new_a7(key_str);
 		atr_itm.Key_rng_(0, key_bry.length);
 		Tfds.Eq(expd, whitelist_mgr.Chk(tag_id, key_bry, atr_itm), key_str);
 	}	private Xop_xatr_itm atr_itm = new Xop_xatr_itm(0, 0);
-	public void Whitelist(byte tag_id, String key_str, String val_str, boolean expd) {
+	public void Whitelist(int tag_id, String key_str, String val_str, boolean expd) {
 		byte[] key_bry = Bry_.new_a7(key_str);
 		atr_itm.Key_rng_(0, key_bry.length);
 		atr_itm.Val_bry_(Bry_.new_a7(val_str));
